Use only stdlib in get_dataset_names
